% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap_aes.R
\name{remap_geom_aes}
\alias{remap_geom_aes}
\title{Remap existing aesthetic mapping in the geom of a layer}
\usage{
remap_geom_aes(layer, mapping, new_aes = NULL)
}
\arguments{
\item{layer}{The layer on which to work}

\item{mapping}{Aesthetic mapping from plot aesthetics to geom aesthetics}

\item{new_aes}{Character vector of the names of any new aesthetics this geom can now handle,
needs to be provided for proper legend drawing}
}
\description{
Remap existing aesthetic mapping in the geom of a layer
}
\examples{
library(ggplot2)
library(dplyr)

df1 <- data.frame(x = c(1, 2, 3))

ggplot(df1, aes(x = x)) +
  geom_tile(aes(fill = x, y = 2)) +
  geom_tile(aes(fill2 = x, y = 1)) \%>\% remap_geom_aes(aes(fill = fill2), new_aes = "fill2") +
  scale_colour_viridis_c(aesthetics = "fill", guide = "legend", name = "viridis A", option = "A") +
  scale_colour_viridis_c(aesthetics = "fill2", guide = "legend", name = "viridis D")


dist_scale <- function(aesthetics, palette, name) {
  scale_colour_distiller(
    aesthetics = aesthetics,
    palette = palette,
    name = name,
    guide = guide_legend(order = palette))
}

ggplot(mapping = aes(Sepal.Length, Sepal.Width)) +
  (geom_point(data = filter(iris, Species == "setosa"), aes(colour1 = Sepal.Width)) \%>\%
     remap_geom_aes(aes(colour = colour1), new_aes = "colour1")) +
  (geom_point(data = filter(iris, Species == "versicolor"), aes(colour2 = Sepal.Width)) \%>\%
     remap_geom_aes(aes(colour = colour2), new_aes = "colour2")) +
  (geom_point(data = filter(iris, Species == "virginica"), aes(colour3 = Sepal.Width)) \%>\%
     remap_geom_aes(aes(colour = colour3), new_aes = "colour3")) +
  facet_wrap(~Species) +
  dist_scale("colour1", 5, "setosa") +
  dist_scale("colour2", 6, "versicolor") +
  dist_scale("colour3", 7, "virginica") +
  theme_bw() +
  theme(legend.position = "bottom", legend.key.width = grid::unit(6, "pt"))

}
