% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remap_aes.R
\name{remap_geom_aes}
\alias{remap_geom_aes}
\title{Remap existing aesthetic mapping in the geom of a layer}
\usage{
remap_geom_aes(layer, old_aes, new_aes)
}
\arguments{
\item{layer}{The layer on which to work}

\item{old_aes}{The old aes (which the geom uses internally)}

\item{new_aes}{The new aes (which we now want the geom to use)}
}
\description{
Remap existing aesthetic mapping in the geom of a layer
}
\examples{
library(ggplot2)
library(dplyr)

df1 <- data.frame(x = c(1, 2, 3))

ggplot(df1, aes(x = x)) +
  geom_tile(aes(fill = x, y = 2)) +
  geom_tile(aes(fill.alt = x, y = 1)) \%>\% remap_geom_aes("fill", "fill.alt") +
  scale_fill_viridis_c(option = "A", guide = "legend", name = "viridis A") +
  scale_colour_viridis_c(aesthetics = "fill.alt", guide = "legend", name = "viridis D")
}
